backend/requirements.txt
# Web framework
fastapi==0.135.2
uvicorn[standard]==0.42.0
python-multipart==0.0.22

# Database
sqlalchemy[asyncio]==2.0.48
asyncpg==0.31.0
alembic==1.18.4

# Auth
authlib==1.6.9
python-jose[cryptography]==3.5.0
itsdangerous==2.2.0
httpx==0.28.1

# Encryption
cryptography==46.0.6

# Google
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1

# LangChain + LangGraph
langchain==1.2.13
langchain-openrouter==0.2.1
langgraph==1.1.3
langgraph-checkpoint-postgres==3.0.5
psycopg[binary,pool]==3.3.3

# Web search
tavily-python==0.7.23

# Config
pydantic-settings==2.13.1
python-dotenv==1.2.2

